Optimizing the Tree
Using various methods to optimize the tree, as well as trying to
delete unused nodes. since the tree is passed in as a parameter,
deleting at the end of the function seems to be interfering with
the tree before the recursive call is complete. still looking
for solutions to this issue.