Part 1
Goal of part one was to create a population for a knapsack
problem, and rank them. Currently this program generates a
population with random attributes, and sorts them within a
population object, where the first stored candidate is the
highest-ranked based on fitness. It prints out the fitnesses
of the population in order, as well as the top-ranked member
data.