the next phase
Just finished writing the code for the next part of this project,
getting the algorithm to select two parents and create a child,
and insert the child if valid. I moved a large amount of
functionality into the Population class, to keep abstraction. I
did get a bug that wouldoccaisionally make a candidate parent be
NaN, but it then is treated as 0, picking the most fit parent.
this could lead to both parents being the top parent, even though
i have a check in place, I haven't seen that bug after testing,
but I'm not confident that it is gone.