Phase 3 Complete
Implemented the Genitor solution, where we accept the solution
after 3 identical homogeneous populations. I ran into an issue
where the population became saturated with a solution but
still had one better solution that couldn't create a better child
using the new, simplified crossover technique. To solve this, I
simply force the population to be homogeneous with that top
solution after a set number of tries (10000 seemed to be much
greater than any attempt that did not have this issue)