Fin
Fixed the knapsack generator(s); apparantly /r/n must be
explicitly typed for the text file to be read correctly.
I had tried changing the write mode in php to add /r
automatically with no success, but it appears that it was
indeed the issue.

on the GA, I reformat some output as well as insert a chance
for mutation in every child (5% chance). It runs reasonably well
with a 100 item knapsack