More Tweaks
Had to output the number of possibilities that would have been used
by an exhaustive search, and it was acceptable to stop calculating
the number where a long long int would overflow. Also, as per a
comment that was left, I added a line to clear my stringstream each
time I get a new line, to keep it in a good state for reading again.
I also fixed the knapGen ruby version to no longer create a list
with one more than was asked for.
Also, a random mutation chance was added, where a child has a 5%
chance to mutate and flip one bit.