Final Submission
Created a different mutate function to just swap the location of
two different pieces, and increased the likelihood of this
mutation occurring. Also, the output is endless, as it never
reaches a fitness of 64 (so it doesn't solve the cube). I did
limit the amount of cout statements that occur to see the curent
fitness.