Reinforcement
So, this update is a mostly finished reinforcement learning AI
that learns TicTacToe. There are some text inputs that aren't
working such as saving and loading training data, exiting the
program, and saturation, but it trains. It doesn't explore
much right now, it likes to get into ruts. At least it learned
something!